+++ src/Scope.php
@@ -14,6 +14,7 @@ namespace League\Fractal;
 use InvalidArgumentException;
 use League\Fractal\Resource\Collection;
 use League\Fractal\Resource\Item;
+use League\Fractal\Resource\Primitive;
 use League\Fractal\Resource\NullResource;
 use League\Fractal\Resource\ResourceInterface;
 use League\Fractal\Serializer\SerializerAbstract;
@@ -258,7 +259,7 @@ class Scope
                 );
             }
 
-            $data = array_merge($data, $includedData);
+            $data = $data + $includedData;
         }
 
         if (!empty($this->availableIncludes)) {
@@ -288,17 +289,47 @@ class Scope
             return null;
         }
 
-        return array_merge($data, $meta);
+        return $data + $meta;
     }
 
     /**
      * Convert the current data for this scope to JSON.
      *
+     * @param int $options
+     *
      * @return string
      */
-    public function toJson()
+    public function toJson($options = 0)
     {
-        return json_encode($this->toArray());
+        return json_encode($this->toArray(), $options);
+    }
+
+    /**
+     * Transformer a primitive resource
+     *
+     * @return mixed
+     */
+    public function transformPrimitiveResource()
+    {
+        if (! ($this->resource instanceof Primitive)) {
+            throw new InvalidArgumentException(
+                'Argument $resource should be an instance of League\Fractal\Resource\Primitive'
+            );
+        }
+
+        $transformer = $this->resource->getTransformer();
+        $data = $this->resource->getData();
+
+        if (null === $transformer) {
+            $transformedData = $data;
+        } elseif (is_callable($transformer)) {
+            $transformedData = call_user_func($transformer, $data);
+        } else {
+            $transformer->setCurrentScope($this);
+            $transformedData = $transformer->transform($data);
+        }
+
+        return $transformedData;
     }
 
     /**