+++ test/Pagination/PhalconFrameworkPaginatorAdapterTest.php
@@ -3,44 +3,29 @@
 namespace League\Fractal\Test\Pagination;
 
 use League\Fractal\Pagination\PhalconFrameworkPaginatorAdapter;
-use Mockery;
+use PHPUnit\Framework\TestCase;
 
-class PhalconFrameworkPaginatorAdapterTest extends \PHPUnit_Framework_TestCase
+class PhalconFrameworkPaginatorAdapterTest extends TestCase
 {
     public function testPaginationAdapter()
     {
-        $total = 50;
-        $count = 10;
-        $perPage = 10;
-        $currentPage = 2;
-        $lastPage = 5;
+        $resultset = new \stdClass();
+        $resultset->items       = array_fill(1, 10, 'fractal');
+        $resultset->current     = 3;
+        $resultset->first       = 1;
+        $resultset->last        = 5;
+        $resultset->next        = 4;
+        $resultset->previous    = 2;
+        $resultset->total_items = 50;
+        $resultset->total_pages = 10;
 
-        $paginate =[
-            'last'        => $lastPage,
-            'current'     => $currentPage,
-            'total_items' => $total,
-            'total_pages' => $count,
-
-        ];
-
-        $paginator = Mockery::mock('Phalcon\Paginator\Adapter\QueryBuilder');
-        $paginator->shouldReceive('currentPage')->andReturn($currentPage);
-        $paginator->shouldReceive('lastPage')->andReturn($lastPage);
-        $paginator->shouldReceive('count')->andReturn($count);
-        $paginator->shouldReceive('total')->andReturn($total);
-        $paginator->shouldReceive('getPaginate')->andReturn((object) $paginate);
-
-        $adapter = new PhalconFrameworkPaginatorAdapter($paginator);
-
+        $adapter = new PhalconFrameworkPaginatorAdapter($resultset);
         $this->assertInstanceOf('League\Fractal\Pagination\PaginatorInterface', $adapter);
-        $this->assertSame($currentPage, $adapter->getCurrentPage());
-        $this->assertSame($lastPage, $adapter->getLastPage());
-        $this->assertSame($count, $adapter->getCount());
-        $this->assertSame($total, $adapter->getTotal());
-    }
-
-    public function tearDown()
-    {
-        Mockery::close();
+        $this->assertSame(3, $adapter->getCurrentPage());
+        $this->assertSame(10, $adapter->getCount());
+        $this->assertSame(50, $adapter->getTotal());
+        $this->assertSame(10, $adapter->getPerPage());
+        $this->assertSame(5, $adapter->getLastPage());
+        $this->assertSame(4, $adapter->getNext());
     }
 }