+++ test/Stub/Transformer/JsonApiEmptyTransformer.php
@@ -1,0 +1,11 @@
+<?php namespace League\Fractal\Test\Stub\Transformer;
+
+use League\Fractal\TransformerAbstract;
+
+class JsonApiEmptyTransformer extends TransformerAbstract
+{
+    public function transform(array $resource)
+    {
+        return $resource;
+    }
+}